small fix + code style improvements
