+++ src/test/java/org/itmo/eventApp/main/controller/EmailSendingTest.java
@@ -1,13 +1,12 @@
 package org.itmo.eventApp.main.controller;
 
-import org.itmo.eventapp.main.mail.MailSenderService;
-import org.junit.Before;
+import org.itmo.eventapp.main.mail.MailSenderServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 
 public class EmailSendingTest extends AbstractTestContainers{
     @Autowired
-    private MailSenderService mailSenderService;
+    private MailSenderServiceImpl mailSenderServiceImpl;
 
-    private GreenMailExtension
+//    private GreenMailExtension
 
 }
